Add TaskPolicy

+++ app/Http/Controllers/TaskController.php
@@ -67,6 +67,8 @@ class TaskController extends Controller
      */
     public function destroy(Request $request, Task $task)
     {
+        $this->authorize('destroy', $task);
+        
         $task->delete();
     
         return redirect('/');
